==> notebooks/0.0-example.ipynb <==
# Example Notebook

A simple example showing you how to cleanly refactor re-usable code.

Don't copy/paste the same functions across notebooks! If you identify a function or class that's useful in multiple places, refactor it out into `src` and then just import it.

import sys
# Insert the parent path relative to this notebook so we can import from the src folder.
sys.path.insert(0, "..")

from src.data.make_dataset import generate_data

dataset = generate_data(length=10)
print(dataset)
