==> notebooks/1.0-web-scrapped-sa-municipalities-demographic.ipynb <==
   ## Scrapping data from https://municipalities.co.za 

#Importing libraries 

from bs4 import BeautifulSoup
import requests
from csv import writer
import pandas as pd
import numpy as np
import boto3

client = boto3.client('s3', aws_access_key_id='AKIATNJHRXAPUA4DIFER', aws_secret_access_key="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")
bucket = "2207-17-fibre-competitive-intensity-model-b"

#### Urls of Local municipalities of South Africa as found on 'https://municipalities.co.za' as at April, 2023


local_urls = ['https://municipalities.co.za/demographic/1181/kheis-local-municipality',
'https://municipalities.co.za/demographic/1112/abaqulusi-local-municipality',
'https://municipalities.co.za/demographic/1238/alfred-duma-local-municipality',
'https://municipalities.co.za/demographic/1003/amahlathi-local-municipality',
'https://municipalities.co.za/demographic/1127/ba-phalaborwa-local-municipality',
'https://municipalities.co.za/demographic/1212/beaufort-west-local-municipality',
'https://municipalities.co.za/demographic/1136/bela-bela-local-municipality',
'https://municipalities.co.za/demographic/1226/bergrivier-local-municipality',
'https://municipalities.co.za/demographic/1239/big-5-hlabisa-local-municipality',
'https://municipalities.co.za/demographic/1215/bitou-local-municipality',
'https://municipalities.co.za/demographic/1118/blouberg-local-municipality',
'https://municipalities.co.za/demographic/1012/blue-crane-route-local-municipality',
'https://municipalities.co.za/demographic/1207/breede-valley-local-municipality',
'https://municipalities.co.za/demographic/1142/bushbuckridge-local-municipality',
'https://municipalities.co.za/demographic/1222/cape-agulhas-local-municipality',
'https://municipalities.co.za/demographic/1227/cederberg-local-municipality',
'https://municipalities.co.za/demographic/1147/chief-albert-luthuli-local-municipality',
'https://municipalities.co.za/demographic/1193/city-of-matlosana-local-municipality',
'https://municipalities.co.za/demographic/1244/city-of-mbombela-local-municipality',
'https://municipalities.co.za/demographic/1110/city-of-umhlathuze-local-municipality',
'https://municipalities.co.za/demographic/1241/collins-chabane-local-municipality',
'https://municipalities.co.za/demographic/1067/dannhauser-local-municipality',
'https://municipalities.co.za/demographic/1245/dawid-kruiper-local-municipality',
'https://municipalities.co.za/demographic/1231/dihlabeng-local-municipality',
'https://municipalities.co.za/demographic/1160/dikgatlong-local-municipality',
'https://municipalities.co.za/demographic/1148/dipaleseng-local-municipality',
'https://municipalities.co.za/demographic/1202/ditsobotla-local-municipality',
'https://municipalities.co.za/demographic/1021/dr-ab-xuma-local-municipality',
'https://municipalities.co.za/demographic/1013/dr-beyers-naude-local-municipality',
'https://municipalities.co.za/demographic/1155/dr-js-moroka-local-municipality',
'https://municipalities.co.za/demographic/1240/dr-nkosazana-dlamini-zuma-local-municipality',
'https://municipalities.co.za/demographic/1153/dr-pixley-ka-isaka-seme-local-municipality',
'https://municipalities.co.za/demographic/1208/drakenstein-local-municipality',
'https://municipalities.co.za/demographic/1113/edumbe-local-municipality',
'https://municipalities.co.za/demographic/1122/elias-motsoaledi-local-municipality',
'https://municipalities.co.za/demographic/1028/elundini-local-municipality',
'https://municipalities.co.za/demographic/1068/emadlangeni-local-municipality',
'https://municipalities.co.za/demographic/1156/emakhazeni-local-municipality',
'https://municipalities.co.za/demographic/1157/emalahleni-local-municipality',
'https://municipalities.co.za/demographic/1020/emalahleni-local-municipality',
'https://municipalities.co.za/demographic/1060/emfuleni-local-municipality',
'https://municipalities.co.za/demographic/1173/emthanjeni-local-municipality',
'https://municipalities.co.za/demographic/1097/endumeni-local-municipality',
'https://municipalities.co.za/demographic/1234/enoch-mgijima-local-municipality',
'https://municipalities.co.za/demographic/1123/ephraim-mogale-local-municipality',
'https://municipalities.co.za/demographic/1243/fetakgomo-tubatse-local-municipality',
'https://municipalities.co.za/demographic/1165/ga-segonyana-local-municipality',
'https://municipalities.co.za/demographic/1164/gamagara-local-municipality',
'https://municipalities.co.za/demographic/1216/george-local-municipality',
'https://municipalities.co.za/demographic/1149/govan-mbeki-local-municipality',
'https://municipalities.co.za/demographic/1005/great-kei-local-municipality',
'https://municipalities.co.za/demographic/1128/greater-giyani-local-municipality',
'https://municipalities.co.za/demographic/1074/greater-kokstad-local-municipality',
'https://municipalities.co.za/demographic/1129/greater-letaba-local-municipality',
'https://municipalities.co.za/demographic/1196/greater-taung-local-municipality',
'https://municipalities.co.za/demographic/1130/greater-tzaneen-local-municipality',
'https://municipalities.co.za/demographic/1167/hantam-local-municipality',
'https://municipalities.co.za/demographic/1217/hessequa-local-municipality',
'https://municipalities.co.za/demographic/1085/impendle-local-municipality',
'https://municipalities.co.za/demographic/1237/inkosi-langalibalele-local-municipality',
'https://municipalities.co.za/demographic/1035/ingquza-hill-local-municipality',
'https://municipalities.co.za/demographic/1023/intsika-yethu-local-municipality',
'https://municipalities.co.za/demographic/1024/inxuba-yethemba-local-municipality',
'https://municipalities.co.za/demographic/1246/jb-marks-local-municipality',
'https://municipalities.co.za/demographic/1166/joe-morolong-local-municipality',
'https://municipalities.co.za/demographic/1093/jozini-local-municipality',
'https://municipalities.co.za/demographic/1232/kagisano-molopo-local-municipality',
'https://municipalities.co.za/demographic/1183/kai-garib-local-municipality',
'https://municipalities.co.za/demographic/1168/kamiesberg-local-municipality',
'https://municipalities.co.za/demographic/1218/kannaland-local-municipality',
'https://municipalities.co.za/demographic/1174/kareeberg-local-municipality',
'https://municipalities.co.za/demographic/1169/karoo-hoogland-local-municipality',
'https://municipalities.co.za/demographic/1184/kgatelopele-local-municipality',
'https://municipalities.co.za/demographic/1187/kgetlengrivier-local-municipality',
'https://municipalities.co.za/demographic/1170/khai-ma-local-municipality',
'https://municipalities.co.za/demographic/1032/king-sabata-dalindyebo-local-municipality',
'https://municipalities.co.za/demographic/1219/knysna-local-municipality',
'https://municipalities.co.za/demographic/1055/kopanong-local-municipality',
'https://municipalities.co.za/demographic/1015/kouga-local-municipality',
'https://municipalities.co.za/demographic/1016/koukamma-local-municipality',
'https://municipalities.co.za/demographic/1070/kwadukuza-local-municipality',
'https://municipalities.co.za/demographic/1213/laingsburg-local-municipality',
'https://municipalities.co.za/demographic/1209/langeberg-local-municipality',
'https://municipalities.co.za/demographic/1150/lekwa-local-municipality',
'https://municipalities.co.za/demographic/1198/lekwa-teemane-local-municipality',
'https://municipalities.co.za/demographic/1119/lepelle-nkumpi-local-municipality',
'https://municipalities.co.za/demographic/1137/lephalale-local-municipality',
'https://municipalities.co.za/demographic/1061/lesedi-local-municipality',
'https://municipalities.co.za/demographic/1056/letsemeng-local-municipality',
'https://municipalities.co.za/demographic/1188/madibeng-local-municipality',
'https://municipalities.co.za/demographic/1039/mafube-local-municipality',
'https://municipalities.co.za/demographic/1161/magareng-local-municipality',
'https://municipalities.co.za/demographic/1203/mahikeng-local-municipality',
'https://municipalities.co.za/demographic/1017/makana-local-municipality',
'https://municipalities.co.za/demographic/1132/makhado-local-municipality',
'https://municipalities.co.za/demographic/1126/makhuduthamaga-local-municipality',
'https://municipalities.co.za/demographic/1051/maluti-a-phofung-local-municipality',
'https://municipalities.co.za/demographic/1199/mamusa-local-municipality',
'https://municipalities.co.za/demographic/1071/mandeni-local-municipality',
'https://municipalities.co.za/demographic/1049/mantsopa-local-municipality',
'https://municipalities.co.za/demographic/1072/maphumulo-local-municipality',
'https://municipalities.co.za/demographic/1192/maquassi-hills-local-municipality',
'https://municipalities.co.za/demographic/1131/maruleng-local-municipality',
'https://municipalities.co.za/demographic/1043/masilonyana-local-municipality',
'https://municipalities.co.za/demographic/1001/matatiele-local-municipality',
'https://municipalities.co.za/demographic/1044/matjhabeng-local-municipality',
'https://municipalities.co.za/demographic/1228/matzikama-local-municipality',
'https://municipalities.co.za/demographic/1006/mbhashe-local-municipality',
'https://municipalities.co.za/demographic/1063/merafong-city-local-municipality',
'https://municipalities.co.za/demographic/1040/metsimaholo-local-municipality',
'https://municipalities.co.za/demographic/1034/mhlontlo-local-municipality',
'https://municipalities.co.za/demographic/1062/midvaal-local-municipality',
'https://municipalities.co.za/demographic/1086/mkhambathini-local-municipality',
'https://municipalities.co.za/demographic/1151/mkhondo-local-municipality',
'https://municipalities.co.za/demographic/1007/mnquma-local-municipality',
'https://municipalities.co.za/demographic/1242/modimolle-mookgophong-local-municipality',
'https://municipalities.co.za/demographic/1139/mogalakwena-local-municipality',
'https://municipalities.co.za/demographic/1064/mogale-city-local-municipality',
'https://municipalities.co.za/demographic/1057/mohokare-local-municipality',
'https://municipalities.co.za/demographic/1120/molemole-local-municipality',
'https://municipalities.co.za/demographic/1041/moqhaka-local-municipality',
'https://municipalities.co.za/demographic/1189/moretele-local-municipality',
'https://municipalities.co.za/demographic/1190/moses-kotane-local-municipality',
'https://municipalities.co.za/demographic/1220/mossel-bay-local-municipality',
'https://municipalities.co.za/demographic/1087/mpofana-local-municipality',
'https://municipalities.co.za/demographic/1152/msukaligwa-local-municipality',
'https://municipalities.co.za/demographic/1088/msunduzi-local-municipality',
'https://municipalities.co.za/demographic/1107/mthonjaneni-local-municipality',
'https://municipalities.co.za/demographic/1107/mthonjaneni-local-municipality',
'https://municipalities.co.za/demographic/1094/mtubatuba-local-municipality',
'https://municipalities.co.za/demographic/1133/musina-local-municipality',
'https://municipalities.co.za/demographic/1045/nala-local-municipality',
'https://municipalities.co.za/demographic/1201/naledi-local-municipality',
'https://municipalities.co.za/demographic/1171/nama-khoi-local-municipality',
'https://municipalities.co.za/demographic/1018/ndlambe-local-municipality',
'https://municipalities.co.za/demographic/1073/ndwedwe-local-municipality',
'https://municipalities.co.za/demographic/1069/newcastle-local-municipality',
'https://municipalities.co.za/demographic/1008/ngqushwa-local-municipality',
'https://municipalities.co.za/demographic/1042/ngwathe-local-municipality',
'https://municipalities.co.za/demographic/1108/nkandla-local-municipality',
'https://municipalities.co.za/demographic/1052/nketoana-local-municipality',
'https://municipalities.co.za/demographic/1144/nkomazi-local-municipality',
'https://municipalities.co.za/demographic/1114/nongoma-local-municipality',
'https://municipalities.co.za/demographic/1099/nquthu-local-municipality',
'https://municipalities.co.za/demographic/1036/ntabankulu-local-municipality',
'https://municipalities.co.za/demographic/1037/nyandeni-local-municipality',
'https://municipalities.co.za/demographic/1104/okhahlamba-local-municipality',
'https://municipalities.co.za/demographic/1221/oudtshoorn-local-municipality',
'https://municipalities.co.za/demographic/1223/overstrand-local-municipality',
'https://municipalities.co.za/demographic/1162/phokwane-local-municipality',
'https://municipalities.co.za/demographic/1053/phumelela-local-municipality',
'https://municipalities.co.za/demographic/1121/polokwane-local-municipality',
'https://municipalities.co.za/demographic/1038/port-st-johns-local-municipality',
'https://municipalities.co.za/demographic/1214/prince-albert-local-municipality',
'https://municipalities.co.za/demographic/1204/ramotshere-moiloa-local-municipality',
'https://municipalities.co.za/demographic/1236/rand-west-city-local-municipality',
'https://municipalities.co.za/demographic/1205/ratlou-local-municipality',
'https://municipalities.co.za/demographic/1080/ray-nkonyeni-local-municipality',
'https://municipalities.co.za/demographic/1233/raymond-mhlaba-local-municipality',
'https://municipalities.co.za/demographic/1175/renosterberg-local-municipality',
'https://municipalities.co.za/demographic/1089/richmond-local-municipality',
'https://municipalities.co.za/demographic/1172/richtersveld-local-municipality',
'https://municipalities.co.za/demographic/1191/rustenburg-local-municipality',
'https://municipalities.co.za/demographic/1026/sakhisizwe-local-municipality',
'https://municipalities.co.za/demographic/1229/saldanha-bay-local-municipality',
'https://municipalities.co.za/demographic/1031/senqu-local-municipality',
'https://municipalities.co.za/demographic/1054/setsoto-local-municipality',
'https://municipalities.co.za/demographic/1176/siyancuma-local-municipality',
'https://municipalities.co.za/demographic/1177/siyathemba-local-municipality',
'https://municipalities.co.za/demographic/1163/sol-plaatje-local-municipality',
'https://municipalities.co.za/demographic/1210/stellenbosch-local-municipality',
'https://municipalities.co.za/demographic/1158/steve-tshwete-local-municipality',
'https://municipalities.co.za/demographic/1019/sundays-river-valley-local-municipality',
'https://municipalities.co.za/demographic/1230/swartland-local-municipality',
'https://municipalities.co.za/demographic/1224/swellendam-local-municipality',
'https://municipalities.co.za/demographic/1145/thaba-chweu-local-municipality',
'https://municipalities.co.za/demographic/1141/thabazimbi-local-municipality',
'https://municipalities.co.za/demographic/1225/theewaterskloof-local-municipality',
'https://municipalities.co.za/demographic/1178/thembelihle-local-municipality',
'https://municipalities.co.za/demographic/1159/thembisile-hani-local-municipality',
'https://municipalities.co.za/demographic/1135/thulamela-local-municipality',
'https://municipalities.co.za/demographic/1046/tokologo-local-municipality',
'https://municipalities.co.za/demographic/1186/tsantsabane-local-municipality',
'https://municipalities.co.za/demographic/1206/tswaing-local-municipality',
'https://municipalities.co.za/demographic/1047/tswelopele-local-municipality',
'https://municipalities.co.za/demographic/1077/ubuhlebezwe-local-municipality',
'https://municipalities.co.za/demographic/1179/ubuntu-local-municipality',
'https://municipalities.co.za/demographic/1115/ulundi-local-municipality',
'https://municipalities.co.za/demographic/1081/umdoni-local-municipality',
'https://municipalities.co.za/demographic/1106/umfolozi-local-municipality',
'https://municipalities.co.za/demographic/1096/umhlabuyalingana-local-municipality',
'https://municipalities.co.za/demographic/1111/umlalazi-local-municipality',
'https://municipalities.co.za/demographic/1090/umngeni-local-municipality',
'https://municipalities.co.za/demographic/1091/umshwathi-local-municipality',
'https://municipalities.co.za/demographic/1098/umsinga-local-municipality',
'https://municipalities.co.za/demographic/1180/umsobomvu-local-municipality',
'https://municipalities.co.za/demographic/1082/umuziwabantu-local-municipality',
'https://municipalities.co.za/demographic/1100/umvoti-local-municipality',
'https://municipalities.co.za/demographic/1078/umzimkhulu-local-municipality',
'https://municipalities.co.za/demographic/1002/umzimvubu-local-municipality',
'https://municipalities.co.za/demographic/1083/umzumbe-local-municipality',
'https://municipalities.co.za/demographic/1116/uphongolo-local-municipality',
'https://municipalities.co.za/demographic/1154/victor-khanye-local-municipality',
'https://municipalities.co.za/demographic/1235/walter-sisulu-local-municipality',
'https://municipalities.co.za/demographic/1033/winnie-madikizela-mandela-local-municipality',
'https://municipalities.co.za/demographic/1211/witzenberg-local-municipality']

#### Urls of District municipalities of South Africa as found on 'https://municipalities.co.za' as at April, 2023.

district_urls = ['https://municipalities.co.za/demographic/101/alfred-nzo-district-municipality',
'https://municipalities.co.za/demographic/116/amajuba-district-municipality',
'https://municipalities.co.za/demographic/102/amathole-district-municipality',
'https://municipalities.co.za/demographic/139/bojanala-platinum-district-municipality',
'https://municipalities.co.za/demographic/143/cape-winelands-district-municipality',
'https://municipalities.co.za/demographic/126/capricorn-district-municipality',
'https://municipalities.co.za/demographic/144/central-karoo-district-municipality',
'https://municipalities.co.za/demographic/104/chris-hani-district-municipality',
'https://municipalities.co.za/demographic/140/dr-kenneth-kaunda-district-municipality',
'https://municipalities.co.za/demographic/141/dr-ruth-segomotsi-mompati-district-municipality',
'https://municipalities.co.za/demographic/131/ehlanzeni-district-municipality',
'https://municipalities.co.za/demographic/107/fezile-dabi-district-municipality',
'https://municipalities.co.za/demographic/134/frances-baard-district-municipality',
'https://municipalities.co.za/demographic/145/garden-route-district-municipality',
'https://municipalities.co.za/demographic/132/gert-sibande-district-municipality',
'https://municipalities.co.za/demographic/118/harry-gwala-district-municipality',
'https://municipalities.co.za/demographic/117/ilembe-district-municipality',
'https://municipalities.co.za/demographic/105/joe-gqabi-district-municipality',
'https://municipalities.co.za/demographic/135/john-taolo-gaetsewe-district-municipality',
'https://municipalities.co.za/demographic/124/king-cetshwayo-district-municipality',
'https://municipalities.co.za/demographic/108/lejweleputswa-district-municipality',
'https://municipalities.co.za/demographic/128/mopani-district-municipality',
'https://municipalities.co.za/demographic/136/namakwa-district-municipality',
'https://municipalities.co.za/demographic/133/nkangala-district-municipality',
'https://municipalities.co.za/demographic/106/or-tambo-district-municipality',
'https://municipalities.co.za/demographic/146/overberg-district-municipality',
'https://municipalities.co.za/demographic/137/pixley-ka-seme-district-municipality',
'https://municipalities.co.za/demographic/103/sarah-baartman-district-municipality',
'https://municipalities.co.za/demographic/114/sedibeng-district-municipality',
'https://municipalities.co.za/demographic/127/sekhukhune-district-municipality',
'https://municipalities.co.za/demographic/110/thabo-mofutsanyana-district-municipality',
'https://municipalities.co.za/demographic/119/ugu-district-municipality',
'https://municipalities.co.za/demographic/120/umgungundlovu-district-municipality',
'https://municipalities.co.za/demographic/121/umkhanyakude-district-municipality',
'https://municipalities.co.za/demographic/122/umzinyathi-district-municipality',
'https://municipalities.co.za/demographic/123/uthukela-district-municipality',
'https://municipalities.co.za/demographic/129/vhembe-district-municipality',
'https://municipalities.co.za/demographic/130/waterberg-district-municipality',
'https://municipalities.co.za/demographic/147/west-coast-district-municipality',
'https://municipalities.co.za/demographic/115/west-rand-district-municipality',
'https://municipalities.co.za/demographic/111/xhariep-district-municipality',
'https://municipalities.co.za/demographic/111/xhariep-district-municipality',
'https://municipalities.co.za/demographic/125/zululand-district-municipality']

#### Urls of Metropolitan municipalities of South Africa as found on 'https://municipalities.co.za' as at April, 2023.

metro_urls = ['https://municipalities.co.za/demographic/7/buffalo-city-metropolitan-municipality',
'https://municipalities.co.za/demographic/6/city-of-cape-town-metropolitan-municipality',
'https://municipalities.co.za/demographic/4/city-of-ekurhuleni-metropolitan-municipality',
'https://municipalities.co.za/demographic/2/city-of-johannesburg-metropolitan-municipality',
'https://municipalities.co.za/demographic/3/city-of-tshwane-metropolitan-municipality',
'https://municipalities.co.za/demographic/5/ethekwini-metropolitan-municipality',
'https://municipalities.co.za/demographic/8/mangaung-metropolitan-municipality',
'https://municipalities.co.za/demographic/1/nelson-mandela-bay-metropolitan-municipality']

#### Creating empty Dataframes to store scrapped data.

df1 = pd.DataFrame()
df2 = pd.DataFrame()
df3 = pd.DataFrame()
df4 = pd.DataFrame()
df5 = pd.DataFrame()
df6 = pd.DataFrame()

#### Creating functions that loop through the above urls to retrieve demographic information.

Creating local_municipality_2016 function

def local_municipality_2016(url):
    
    url = url
    page = requests.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    
    #Accessing the table that contains the demographic information from the webpage
    tab = soup.find('table', class_='table table-hover table-bordered table-condensed')  
    title = soup.find('h1', class_='header').text    #Retrieving the name of the municipality
    tr = tab.find_all('tr')                          #Retrieving all the rows of the table

    Column0 = []
    Column1 = [] 
    Column2 = [] 
    
    """The formatting of the first two rows of the table are a bit different than the rest.
    So, data was retieved from these two rows seperately"""

    Column0.append('Municipality')
    row1 = tr[0].find_all('td')
    Column1.append(title + (row1[0].text))
    Column2.append(title + (row1[1].text))

    Column0.append(tr[1].find('th').text)
    row2 = tr[1].find_all('td')
    Column1.append(row2[0].text)
    Column2.append(row2[1].text)
    

    for row in tr:
        if row.find_all('th', 'colspan' == 3):            #Exempting rows that were not needed.
            th = row.find_all('th', 'colspan' == 3)

        else:                                             #Retrieving data from each row of the table
            td = row.find_all('td')
            Column0.append(td[0].text)
            Column1.append(td[1].text)
            Column2.append(td[2].text)
            
    global df1
    
    ser2016 = pd.Series(Column1)                          #Turning the list to a pandas Series
    Row2016 = pd.DataFrame(ser2016.values.reshape(1, 22)) #Reshaping the Series and converting it to a Dataframe


    df1 = pd.concat([df1, Row2016.copy()], axis = 0, ignore_index = True )

Creating local_municipality_2011 function

def local_municipality_2011(url):
    
    url = url
    page = requests.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    
    #Accessing the table that contains the demographic information from the webpage
    tab = soup.find('table', class_='table table-hover table-bordered table-condensed')  
    title = soup.find('h1', class_='header').text    #Retrieving the name of the municipality
    tr = tab.find_all('tr')                          #Retrieving all the rows of the table

    Column0 = []
    Column1 = [] 
    Column2 = [] 
    
    """The formatting of the first two rows of the table are a bit different than the rest.
    So, data was retieved from these two rows seperately"""

    Column0.append('Municipality')
    row1 = tr[0].find_all('td')
    Column1.append(title + (row1[0].text))
    Column2.append(title + (row1[1].text))

    Column0.append(tr[1].find('th').text)
    row2 = tr[1].find_all('td')
    Column1.append(row2[0].text)
    Column2.append(row2[1].text)
    

    for row in tr:
        if row.find_all('th', 'colspan' == 3):            #Exempting rows that were not needed.
            th = row.find_all('th', 'colspan' == 3)

        else:                                             #Retrieving data from each row of the table
            td = row.find_all('td')
            Column0.append(td[0].text)
            Column1.append(td[1].text)
            Column2.append(td[2].text)
            
    global df2
    
    ser2011 = pd.Series(Column2)                          #Turning the list to a pandas Series
    Row2011 = pd.DataFrame(ser2011.values.reshape(1, 22)) #Reshaping the Series and converting it to a Dataframe


    df2 = pd.concat([df2, Row2011.copy()], axis = 0, ignore_index = True )

Creating district_municipality_2016 function

def district_municipality_2016(url):
    
    url = url
    page = requests.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    
    #Accessing the table that contains the demographic information from the webpage
    tab = soup.find('table', class_='table table-hover table-bordered table-condensed')  
    title = soup.find('h1', class_='header').text    #Retrieving the name of the municipality
    tr = tab.find_all('tr')                          #Retrieving all the rows of the table

    Column0 = []
    Column1 = [] 
    Column2 = [] 
    
    """The formatting of the first two rows of the table are a bit different than the rest.
    So, data was retieved from these two rows seperately"""

    Column0.append('Municipality')
    row1 = tr[0].find_all('td')
    Column1.append(title + (row1[0].text))
    Column2.append(title + (row1[1].text))

    Column0.append(tr[1].find('th').text)
    row2 = tr[1].find_all('td')
    Column1.append(row2[0].text)
    Column2.append(row2[1].text)
    

    for row in tr:
        if row.find_all('th', 'colspan' == 3):            #Exempting rows that were not needed.
            th = row.find_all('th', 'colspan' == 3)

        else:                                             #Retrieving data from each row of the table
            td = row.find_all('td')
            Column0.append(td[0].text)
            Column1.append(td[1].text)
            Column2.append(td[2].text)
            
    global df3
    
    ser2016 = pd.Series(Column1)                          #Turning the list to a pandas Series
    Row2016 = pd.DataFrame(ser2016.values.reshape(1, 22)) #Reshaping the Series and converting it to a Dataframe


    df3 = pd.concat([df3, Row2016.copy()], axis = 0, ignore_index = True )

Creating district_municipality_2011 function

def district_municipality_2011(url):
    
    url = url
    page = requests.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    
    #Accessing the table that contains the demographic information from the webpage
    tab = soup.find('table', class_='table table-hover table-bordered table-condensed')  
    title = soup.find('h1', class_='header').text    #Retrieving the name of the municipality
    tr = tab.find_all('tr')                          #Retrieving all the rows of the table

    Column0 = []
    Column1 = [] 
    Column2 = [] 
    
    """The formatting of the first two rows of the table are a bit different than the rest.
    So, data was retieved from these two rows seperately"""

    Column0.append('Municipality')
    row1 = tr[0].find_all('td')
    Column1.append(title + (row1[0].text))
    Column2.append(title + (row1[1].text))

    Column0.append(tr[1].find('th').text)
    row2 = tr[1].find_all('td')
    Column1.append(row2[0].text)
    Column2.append(row2[1].text)
    

    for row in tr:
        if row.find_all('th', 'colspan' == 3):            #Exempting rows that were not needed.
            th = row.find_all('th', 'colspan' == 3)

        else:                                             #Retrieving data from each row of the table
            td = row.find_all('td')
            Column0.append(td[0].text)
            Column1.append(td[1].text)
            Column2.append(td[2].text)
            
    global df4
    
    
    ser2011 = pd.Series(Column2)                           #Turning the list to a pandas Series
    Row2011 = pd.DataFrame(ser2011.values.reshape(1, 22))  #Reshaping the Series and converting it to a Dataframe

  
    df4 = pd.concat([df4, Row2011.copy()], axis = 0, ignore_index = True )

#### Creating metro_municipality_2016 function

def metro_municipality_2016(url):
    
    url = url
    page = requests.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    
    #Accessing the table that contains the demographic information from the webpage
    tab = soup.find('table', class_='table table-hover table-bordered table-condensed')  
    title = soup.find('h1', class_='header').text    #Retrieving the name of the municipality
    tr = tab.find_all('tr')                          #Retrieving all the rows of the table

    Column0 = []
    Column1 = [] 
    Column2 = [] 
    
    """The formatting of the first two rows of the table are a bit different than the rest.
    So, data was retieved from these two rows seperately"""

    Column0.append('Municipality')
    row1 = tr[0].find_all('td')
    Column1.append(title + (row1[0].text))
    Column2.append(title + (row1[1].text))

    Column0.append(tr[1].find('th').text)
    row2 = tr[1].find_all('td')
    Column1.append(row2[0].text)
    Column2.append(row2[1].text)
    

    for row in tr:
        if row.find_all('th', 'colspan' == 3):            #Exempting rows that were not needed.
            th = row.find_all('th', 'colspan' == 3)

        else:                                             #Retrieving data from each row of the table
            td = row.find_all('td')
            Column0.append(td[0].text)
            Column1.append(td[1].text)
            Column2.append(td[2].text)
            
    global df5
    
    
    ser2016 = pd.Series(Column1)                           #Turning the list to a pandas Series
    Row2016 = pd.DataFrame(ser2016.values.reshape(1, 22))  #Reshaping the Series and converting it to a Dataframe

  
    df5 = pd.concat([df5, Row2016.copy()], axis = 0, ignore_index = True )

#### Creating metro_municipality_2011 function

def metro_municipality_2011(url):
    
    url = url
    page = requests.get(url)
    soup = BeautifulSoup(page.content, 'html.parser')
    
    #Accessing the table that contains the demographic information from the webpage
    tab = soup.find('table', class_='table table-hover table-bordered table-condensed')  
    title = soup.find('h1', class_='header').text    #Retrieving the name of the municipality
    tr = tab.find_all('tr')                          #Retrieving all the rows of the table

    Column0 = []
    Column1 = [] 
    Column2 = [] 
    
    """The formatting of the first two rows of the table are a bit different than the rest.
    So, data was retieved from these two rows seperately"""

    Column0.append('Municipality')
    row1 = tr[0].find_all('td')
    Column1.append(title + (row1[0].text))
    Column2.append(title + (row1[1].text))

    Column0.append(tr[1].find('th').text)
    row2 = tr[1].find_all('td')
    Column1.append(row2[0].text)
    Column2.append(row2[1].text)
    

    for row in tr:
        if row.find_all('th', 'colspan' == 3):            #Exempting rows that were not needed.
            th = row.find_all('th', 'colspan' == 3)

        else:                                             #Retrieving data from each row of the table
            td = row.find_all('td')
            Column0.append(td[0].text)
            Column1.append(td[1].text)
            Column2.append(td[2].text)
            
    global df6
    
    
    ser2011 = pd.Series(Column2)                           #Turning the list to a pandas Series
    Row2011 = pd.DataFrame(ser2011.values.reshape(1, 22))  #Reshaping the Series and converting it to a Dataframe

  
    df6 = pd.concat([df6, Row2011.copy()], axis = 0, ignore_index = True )

Calling the local_municipality_2016 function to scrape and concat the 2016 local municipalities data.

for url in local_urls:
    local_municipality_2016(url)

Adding columns to the scrapped 2016 local municipalities data

df1.columns = ['Municipality', 'Population', 'Population under 15', 'Population 15 to 64', 'Population over 65', 'Dependency Ratio: Per 100 (15-64)', 'Sex Ratio: Males per 100 females', 'Population growth Per annum', 'Unemployment rate (official)', 'Youth unemployment rate (official) 15-34', 'No schooling', 'Matric', 'Higher education', 'Households', 'Average household size', 'Female headed households', 'Formal dwellings', 'Housing owned', 'Flush toilet connected to sewerage', 'Weekly refuse removal', 'Piped water inside dwelling', 'Electricity for lighting']

Converting the dataframe to a csv

df1.to_csv("2016-South-Africa-Local-Municipalities.csv")

client.upload_file("2016-South-Africa-Local-Municipalities.csv", Bucket=bucket, Key="Raw Data/2016-South-Africa-Local-Municipalities.csv")

Calling the local_municipality_2011 function to scrape and concat the 2011 local municipalities data.

for url in local_urls:
    local_municipality_2011(url)

Adding columns to the scrapped 2011 local municipalities data.

df2.columns = ['Municipality', 'Population', 'Population under 15', 'Population 15 to 64', 'Population over 65', 'Dependency Ratio: Per 100 (15-64)', 'Sex Ratio: Males per 100 females', 'Population growth Per annum', 'Unemployment rate (official)', 'Youth unemployment rate (official) 15-34', 'No schooling', 'Matric', 'Higher education', 'Households', 'Average household size', 'Female headed households', 'Formal dwellings', 'Housing owned', 'Flush toilet connected to sewerage', 'Weekly refuse removal', 'Piped water inside dwelling', 'Electricity for lighting']

Converting the dataframe to a csv

df2.to_csv("2011-South-Africa-Local-Municipalities.csv")

client.upload_file("2011-South-Africa-Local-Municipalities.csv", Bucket=bucket, Key="Raw Data/2011-South-Africa-Local-Municipalities.csv")

Calling the district_municipality_2016 function to scrape and concat the 2016 district municipalities data.

for url in district_urls:
    district_municipality_2016(url)

Adding columns to the scrapped 2016 district municipalities data.

df3.columns = ['Municipality', 'Population', 'Population under 15', 'Population 15 to 64', 'Population over 65', 'Dependency Ratio: Per 100 (15-64)', 'Sex Ratio: Males per 100 females', 'Population growth Per annum', 'Unemployment rate (official)', 'Youth unemployment rate (official) 15-34', 'No schooling', 'Matric', 'Higher education', 'Households', 'Average household size', 'Female headed households', 'Formal dwellings', 'Housing owned', 'Flush toilet connected to sewerage', 'Weekly refuse removal', 'Piped water inside dwelling', 'Electricity for lighting']

Converting the dataframe to a csv

df3.to_csv("2016-South-Africa-District-Municipalities.csv")

client.upload_file("2016-South-Africa-District-Municipalities.csv", Bucket=bucket, Key="Raw Data/2016-South-Africa-District-Municipalities.csv")

Calling the district_municipality_2011 function to scrape and concat the 2011 district municipalities data.

for url in district_urls:
    district_municipality_2011(url)

Adding columns to the scrapped 2011 district municipalities data.

df4.columns = ['Municipality', 'Population', 'Population under 15', 'Population 15 to 64', 'Population over 65', 'Dependency Ratio: Per 100 (15-64)', 'Sex Ratio: Males per 100 females', 'Population growth Per annum', 'Unemployment rate (official)', 'Youth unemployment rate (official) 15-34', 'No schooling', 'Matric', 'Higher education', 'Households', 'Average household size', 'Female headed households', 'Formal dwellings', 'Housing owned', 'Flush toilet connected to sewerage', 'Weekly refuse removal', 'Piped water inside dwelling', 'Electricity for lighting']

Converting the dataframe to a csv

df4.to_csv("2011-South-Africa-District-Municipalities.csv")

client.upload_file("2011-South-Africa-District-Municipalities.csv", Bucket=bucket, Key="Raw Data/2011-South-Africa-District-Municipalities.csv")

Calling the metro_municipality_2016 function to scrape and concat the 2016 metro municipalities data.

for url in metro_urls:
    metro_municipality_2016(url)

Adding columns to the scrapped 2016 metropolitan municipalities data.

df5.columns = ['Municipality', 'Population', 'Population under 15', 'Population 15 to 64', 'Population over 65', 'Dependency Ratio: Per 100 (15-64)', 'Sex Ratio: Males per 100 females', 'Population growth Per annum', 'Unemployment rate (official)', 'Youth unemployment rate (official) 15-34', 'No schooling', 'Matric', 'Higher education', 'Households', 'Average household size', 'Female headed households', 'Formal dwellings', 'Housing owned', 'Flush toilet connected to sewerage', 'Weekly refuse removal', 'Piped water inside dwelling', 'Electricity for lighting']

Converting the dataframe to a csv

df5.to_csv("2016-South-Africa-Metropolitan-Municipalities.csv")

client.upload_file("2016-South-Africa-Metropolitan-Municipalities.csv", Bucket=bucket, Key="Raw Data/2016-South-Africa-Metropolitan-Municipalities.csv")

Calling the metro_municipality_2011 function to scrape and concat the 2011 metro municipalities data.

for url in metro_urls:
    metro_municipality_2011(url)

Adding columns to the scrapped 2011 metropolitan municipalities data.

df6.columns = ['Municipality', 'Population', 'Population under 15', 'Population 15 to 64', 'Population over 65', 'Dependency Ratio: Per 100 (15-64)', 'Sex Ratio: Males per 100 females', 'Population growth Per annum', 'Unemployment rate (official)', 'Youth unemployment rate (official) 15-34', 'No schooling', 'Matric', 'Higher education', 'Households', 'Average household size', 'Female headed households', 'Formal dwellings', 'Housing owned', 'Flush toilet connected to sewerage', 'Weekly refuse removal', 'Piped water inside dwelling', 'Electricity for lighting']

Converting the dataframe to a csv

df6.to_csv("2011-South-Africa-Metropolitan-Municipalities.csv")

client.upload_file("2011-South-Africa-Metropolitan-Municipalities.csv", Bucket=bucket, Key="Raw Data/2011-South-Africa-Metropolitan-Municipalities.csv")
