==> notebooks/1.1-sa-census-2011-household-data-preprocessing.ipynb <==
import pandas as pd
import numpy as np
import boto3

client = boto3.client('s3', aws_access_key_id='AKIATNJHRXAPUA4DIFER', aws_secret_access_key="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")
bucket = "2207-17-fibre-competitive-intensity-model-b"

df = pd.read_stata("https://2207-17-fibre-competitive-intensity-model-b.s3.eu-west-1.amazonaws.com/Raw+Data/sa-census-2011-household-v1.1-20140618.dta")

pd.set_option('display.max_columns', None)
df.tail(2)

# Changing the income_class column to string to be able to perform string manipulation.
df.DERH_INCOME_CLASS = df.DERH_INCOME_CLASS.astype(str)

# Stripping Rand and spaces from the column
income = df.DERH_INCOME_CLASS.apply(lambda x: x.strip().replace('R', '').replace(' ', ''))

# Stripping hyphen from the column
income = incme.apply(lambda x: x.split('-'))

# Checking the different values in the column
income.value_counts().index

# Creating a column for the lower bounds of the income class
df['Income_Class_Lower']  = b.apply(lambda x: x[0])

# Creating a column for the upper bounds of the income class
df['Income_Class_Higher'] = b.apply(lambda x: x[-1])

# Checking the different values in the column
df['Income_Class_Higher'].value_counts().index

# Replacing '2457601ormore' with '2457601' in the column
df['Income_Class_Higher'] = df['Income_Class_Higher'].apply(lambda x: x.replace('2457601ormore', '2457601'))

# Replacing 'Noincome' with '0' in the column
df['Income_Class_Higher'] = df['Income_Class_Higher'].apply(lambda x: x.replace('Noincome', '0'))

# Replacing 'Unspecified' with '0' in the column
df['Income_Class_Higher'] = df['Income_Class_Higher'].apply(lambda x: x.replace('Unspecified', '0'))

# Checking the different values in the column
df['Income_Class_Lower'].value_counts().index

# Replacing 'Noincome' with '0' in the column
df['Income_Class_Lower'] = df['Income_Class_Lower'].apply(lambda x: x.replace('Noincome', '0'))

# Replacing 'Unspecified' with '0' in the column
df['Income_Class_Lower'] = df['Income_Class_Lower'].apply(lambda x: x.replace('Unspecified', '0'))

# Replacing '2457601ormore' with '2457601' in the column
df['Income_Class_Lower'] = df['Income_Class_Lower'].apply(lambda x: x.replace('2457601ormore', '2457601'))

# Changing the type of the column to integer
df['Income_Class_Higher'] = df['Income_Class_Higher'].astype(int)

# Changing the type of the column to integer
df['Income_Class_Lower'] = df['Income_Class_Lower'].astype(int)

# Creating an average income from Income_Class_Higher and Income_Class_Lower
df['Average_Income'] = df[['Income_Class_Higher', 'Income_Class_Lower']].apply(np.mean, axis = 1)

df.drop(['SN','QN_TYPE'], axis = 1, inplace=True)

df.head(2)

df.to_stata('Cleaned_2011_SA_Census_Sample.dta', compression = 'infer' )

client.upload_file("Cleaned_2011_SA_Census_Sample.dta", Bucket=bucket, Key="Preprocessed Data/Cleaned_2011_SA_Census_Sample.dta")
