==> notebooks/1.5-preprocessed-2016-scraped-municipal-data.ipynb <==
# Preprocess 2016 SCraped Municipalities Data

# import necessary dependecies
import numpy as np
import pandas as pd
import boto3

# Instantiate boto3 by providing access and secrete keys
client = boto3.client('s3', aws_access_key_id='AKIATNJHRXAPUA4DIFER', aws_secret_access_key="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")

# S3 bucket name
bucket = "2207-17-fibre-competitive-intensity-model-b"

# Generate a file path to the S3 bucket
file_path = 'https://2207-17-fibre-competitive-intensity-model-b.s3.eu-west-1.amazonaws.com/Raw+Data/2016-South-Africa-Local-Municipalities.csv'

# Load municipal data
df = pd.read_csv(file_path)

# display the first few rows of the DataFrame
df.head(5)

# Unnamed column is a redundant index
df.drop('Unnamed: 0', axis=1, inplace=True)

# Check for missing rows
df.isnull().sum()

# unemployment rates have zero entries, so the two columns will be dropped
df.drop(['Unemployment rate (official)', 'Youth unemployment rate (official) 15-34'], axis=1, inplace=True)

# Confirm missing rows
df.isnull().sum()

# show a summary of the DataFrame
df.info()

# Convert columns to appropriate data types and convert the percentage values to decimal values
df['Population'] = df['Population'].replace('[^0-9]', '', regex=True).astype(int)
df['Population under 15'] = df['Population under 15'].replace('[^0-9.]', '', regex=True).astype(float)
df['Population 15 to 64'] = df['Population 15 to 64'].replace('[^0-9.]', '', regex=True).astype(float)
df['Population over 65'] = df['Population over 65'].replace('[^0-9.]', '', regex=True).astype(float)
df['Dependency Ratio: Per 100 (15-64)'] = df['Dependency Ratio: Per 100 (15-64)'].astype(float)
df['Sex Ratio: Males per 100 females'] = df['Sex Ratio: Males per 100 females'].astype(float)
df['No schooling'] = df['No schooling'].replace('[^0-9.]', '', regex=True).astype(float)
df['Matric'] = df['Matric'].replace('[^0-9.]', '', regex=True).astype(float)
df['Higher education'] = df['Higher education'].replace('[^0-9.]', '', regex=True).astype(float)
df['Households'] = df['Households'].replace('[^0-9]', '', regex=True).astype(int)
df['Average household size'] = df['Average household size'].astype(float)
df['Female headed households'] = df['Female headed households'].replace('[^0-9.]', '', regex=True).astype(float)
df['Formal dwellings'] = df['Formal dwellings'].replace('[^0-9.]', '', regex=True).astype(float)
df['Housing owned'] = df['Housing owned'].replace('[^0-9.]', '', regex=True).astype(float)
df['Flush toilet connected to sewerage'] = df['Flush toilet connected to sewerage'].replace('[^0-9.]', '', regex=True).astype(float)
df['Weekly refuse removal'] = df['Weekly refuse removal'].replace('[^0-9.]', '', regex=True).astype(float)
df['Piped water inside dwelling'] = df['Piped water inside dwelling'].replace('[^0-9.]', '', regex=True).astype(float)
df['Electricity for lighting'] = df['Electricity for lighting'].replace('[^0-9.]', '', regex=True).astype(float)

# show a summary of the DataFrame to confirm dtype conversion
df.info()

# display the first few rows of the DataFrame to confirm percentage to decimal conversion
df.head()

# Check for duplicates
df.duplicated().sum()


# show duplicated rows
df[df.duplicated(keep=False)]

# keep the first and drop other duplicated rows
df.drop_duplicates(keep='first', inplace=True)

# check for duplicates to confirm duplicated row has been dropped
df.duplicated().sum()

# rename the columns appropriately. 
# I added "percent_" to the beginning of each column name that represents a percentage of the population.

df = df.rename(columns={
    'Municipality': 'municipality',
    'Population': 'population',
    'Population under 15': 'percent_population_under_15',
    'Population 15 to 64': 'percent_population_15_to_64',
    'Population over 65': 'percent_population_over_65',
    'Dependency Ratio: Per 100 (15-64)': 'dependency ratio:per_100_15_to_64',
    'Sex Ratio: Males per 100 females': 'sex ratio: males per 100 females',
    'No schooling': 'percent_no_schooling',
    'Matric': 'percent_matric',
    'Higher education': 'percent_higher_education',
    'Households': 'households',
    'Average household size': 'average_household_size',
    'Female headed households': 'percent_female_headed_households',
    'Formal dwellings': 'percent_formal_dwellings',
    'Housing owned': 'percent_housing_owned',
    'Flush toilet connected to sewerage': 'percent_flush_toilet_connected_to_sewerage',
    'Weekly refuse removal': 'percent_weekly_refuse_removal',
    'Piped water inside dwelling': 'percent_piped_water_inside_dwelling',
    'Electricity for lighting': 'percent_electricity_for_lighting'
})



Emalahleni municipality exist for two different districts and these municipalities are valid. to avoid any problem when annlyzing, we will attach their district names to them

# Create a list with the old and new names for the municipalities
old_name = ['Emalahleni Local Municipality (MP312)2016', 'Emalahleni Local Municipality (EC136)2016']
new_name = ['Emalahleni Nkangala Local Municipality (MP312)2016', 'Emalahleni Chris Hani Local Municipality (EC136)2016']

# Rename the municipalities
df['municipality'] = df['municipality'].replace(old_name, new_name)

# strip municipality name strings to municipality names only
df['municipality'] = df['municipality'].apply(lambda x: x[:x.index('Local')]).str.strip()

# display the first few rows of the DataFrame to confirm the changes in the two steps above
df.head()

# Write the cleaned dataframe to a new CSV file
df.to_csv('cleaned-2016-scraped-local-municipal-data.csv')

# Upload the file to the S3 bucket
client.upload_file("cleaned-2016-scraped-local-municipal-data.csv", Bucket=bucket, 
                   Key="Preprocessed Data/cleaned-2016-scraped-local-municipal-data.csv")
