==> notebooks/1.9-prepare-data-for-municipality-algorithm-development.ipynb <==
# Preprocess Data for Algorithm Development

# Import the relevant libraries
import numpy as np
import pandas as pd
import boto3

# Instantiate boto3 by providing access and secrete keys
client = boto3.client('s3', aws_access_key_id='AKIATNJHRXAPUA4DIFER', aws_secret_access_key="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")

# S3 bucket name
bucket = "2207-17-fibre-competitive-intensity-model-b"

# Generate a file path to the S3 bucket
census_file_path = 'https://2207-17-fibre-competitive-intensity-model-b.s3.eu-west-1.amazonaws.com/Preprocessed+Data/Cleaned_2011_SA_Census_Sample.dta'
uptake_file_path = 'https://2207-17-fibre-competitive-intensity-model-b.s3.eu-west-1.amazonaws.com/Preprocessed+Data/SA-municipality-uptake-rate.csv'

# Load the datasets
df_census = pd.read_stata(census_file_path)
df_uptake = pd.read_csv(uptake_file_path)

# Check the features of the census data
df_census.columns

# Preview the top five rows of the census data
df_census.head()

# Preview the top five rows of the uptake rate dataset
df_uptake.head()

The municpality names in the census data and uptake rate data have some descrepancies as some municipalities have been disestablished in the uptake rate municipality and some municipality names have also changed. There is a need to align the municipalities to share the same name for easy merging of the two dataframes

# Rename the Emalahleni municipalities to their original names so as to be easily merged with the census data
df_uptake.loc[(df_uptake['municipality'] == 'emalahleni') & (df_uptake['DISTRICT_N'] == 'Nkangala'), 'municipality'] = 'emalahleni-mp'
df_uptake.loc[(df_uptake['municipality'] == 'emalahleni') & (df_uptake['DISTRICT_N'] == 'Chris Hani'), 'municipality'] = 'emalahleni-ec'

# Store the names of municipalities to be replaced in a list
old_mun_name = ['emnambithi/ladysmith','indaka','the big 5 false bay','hlabisa','albert luthuli','umjindi','mbombela','mier',
                '//khara hais','engcobo','camdeboo','ikwezi','baviaans','ingwe','kwa sani','pixley ka seme','tsolwana','inkwanca',
               'lukanji','fetakgomo','greater tubatse','ga-segonyane','umtshezi','imbabazane','ventersdorp',
                'tlokwe city council','madibeng','mookgopong','modimolle','naledi-fs','naledi-nw','randfontein','westonaria','ezingoleni',
               'hibiscus coast','nkonkobe','nxuba','sol plaatjie','maletswai','gariep']
# Store the new names of municipalities in a list
new_mun_name = ['alfred duma','alfred duma','big five hlabisa','big five hlabisa','chief albert luthuli','city of mbombela','city of mbombela',
               'dawid kruiper','dawid kruiper','dr ab xuma','dr beyers naude','dr beyers naude','dr beyers naude',
               'dr nkosazana dlamini zuma','dr nkosazana dlamini zuma','dr pixley ka isaka seme','enoch mgijima','enoch mgijima',
               'enoch mgijima','fetakgomo tubatse','fetakgomo tubatse','ga-segonyana','inkosi langalibalele',
                'inkosi langalibalele','jb marks','jb marks','local municipality of madibeng', 'modimolle-mookgophong',
               'modimolle-mookgophong','naledi','naledi','rand west city','rand west city','ray nkonyeni','ray nkonyeni',
                'raymond mhlaba','raymond mhlaba','sol plaatje','walter sisulu','walter sisulu']


# Ensure that the length of both lists are equal
len(new_mun_name) == len(old_mun_name)

# Replace the old municipality names with the new one in the census dataset
df_census['H_MUNIC'] = df_census['H_MUNIC'].str.lower().replace(old_mun_name, new_mun_name)

# Check for municipalities that are in the uptake rate dataset but are not in the census dataset
[i for i in list(df_uptake['municipality']) if i not in list(df_census['H_MUNIC'].str.lower().unique())]

The "Collins Chabane" municipality had no match in the 2011 census data as municipality was created in 2016 from some portions of Thulamela and Makhado municipalities

# Check for municipalities that are in the census dataset but are not in the uptake rate dataset
[i for i in list(df_census['H_MUNIC'].str.lower().unique()) if i not in list(df_uptake['municipality'])]

These three municipalities that were not found in the speed test data were disestablished in 2016 and were either integrated into another municipality or broken into multiple municipalities.

# Extract the columns of interest from the census dataset
df_new_census = df_census.iloc[:, -13:]

## Feature Engineering

# Engineer numerical values from the employment status column
df_new_census['DERH_HH_EMPLOY_STATUS'] = pd.to_numeric(df_new_census['DERH_HH_EMPLOY_STATUS'].map({'Employed': 1,
                                                                                     'Not economically active':0,
                                                                                     'Unemployed':0,
                                                                                     'Household head out of working age scope i.e. ...':0,
                                                                                     'Discouraged work-seeker':0 }))

# Fill the missing values in the employment status
df_new_census['DERH_HH_EMPLOY_STATUS'] = df_new_census['DERH_HH_EMPLOY_STATUS'].fillna(0)

# Map the household geotypes into numerical components of appropriate weights
df_new_census['H_GEOTYPE'] = pd.to_numeric(df_new_census['H_GEOTYPE'].map({'Farms':0.5, 'Urban':1.5,'Traditional':1}))

df_new_census

# Separate the categorical feautures from the numeric features so as to facilitate aggregation
df_census_cat = df_new_census[['H_MUNIC', 'H_GEOTYPE','DERH_HH_EMPLOY_STATUS']]

# Drop features that categorical features to keep only the numeric ones
df_new_census.drop(['DERH_HH_EMPLOY_STATUS','DERH_XPOP','DERH_INCOME_CLASS','HHLD_10PERCENT_WGT',
                    'H_DISTRICT','H_PROVINCE','H_GEOTYPE'], axis=1, inplace=True)

# Aggregate the numeric census dataset and reset the index
grouped_mun_numeric = df_new_census.groupby(['H_MUNIC']).mean().reset_index()

# Aggregate the categorical census dataset and reset the index
grouped_mun_cat = df_census_cat.groupby(['H_MUNIC']).sum().reset_index()

# Combine the two census datasets into one
df_agg_census = pd.merge(grouped_mun_cat,grouped_mun_numeric, on='H_MUNIC')

# View top five rows of the combined datasets
df_agg_census.head()

# Create a combined dataframe for the census and uptake rates
df_cleaned_merge = pd.merge(df_uptake, df_agg_census, how='left', left_on='municipality', right_on='H_MUNIC')

# View the top five rows of the merged datasets to ensure everything went smoothly
df_cleaned_merge.head()

# Check for missing values
df_cleaned_merge.isnull().sum()

# Export dataset for modelling
df_cleaned_merge.to_csv("municipality-data-for-modelling.csv")

# Upload the file to the S3 bucket
client.upload_file("municipality-data-for-modelling.csv", Bucket=bucket, Key="Data for Modeling/municipality-data-for-modelling.csv")
