==> notebooks/2.4-Preprocessing_ward_level_age_data.ipynb <==
#### Import necessary dependencies

import pandas as pd
import boto3

#### Read the CSV file

df = pd.read_excel('https://2207-17-fibre-competitive-intensity-model-b.s3.eu-west-1.amazonaws.com/2011+Census+Raw+Demographic+Data+Ward+Level+/table_2011_age.xlsx', header=0)

#### Display the first few rows of the DataFrame to confirm the changes in the two steps above

df.head()

#### Aggregate the age into different categories


# Calculate the percentage of population less than 15 years
df['percent_less_15'] = ((df['00 - 04'] + df['05 - 09'] + df['10 - 14'])/df['Total'])

# Calculate the percentage of population less than 15 but less than 25 years
df['percent_more_15_less_25'] = ((df['15 - 19'] + df['20 - 24'])/df['Total'])

# Calculate the percentage of population greater than 25 but less than 60 years
df['percent_more_25_less_60'] = ((df['25 - 29'] + df['30 - 34'] + df['35 - 39'] + df['40 - 44'] + df['45 - 49'] + df['50 - 54'] + df['55 - 59'])/df['Total'])

# Calculate the percentage of population greater than 60 years
df['percent_greater_60'] = ((df['60 - 64']+ df['65 - 69'] + df['70 - 74'] + df['75 - 79'] + df['80 - 84'] + df['85+'])/df['Total'])

# Display the modified DataFrame
df.head()


### Rename the columns appropriately

ward_id = df['ward_id']
Total = df['Total']
percent_less_15 = df['percent_less_15']
percent_more_15_less_25 = df['percent_more_15_less_25']
percent_more_25_less_60 = df['percent_more_25_less_60']
percent_greater_60 = df['percent_greater_60']


df_age_aggregated = pd.DataFrame({
    'ward_id' : ward_id,
    'total_pop' : Total,
    'percent_less_15' : percent_less_15,
    'percent_more_15_less_25' : percent_more_15_less_25,    
    'percent_more_25_less_60' : percent_more_25_less_60,
    'percent_greater_60' : percent_greater_60
})

#### Display the first few rows of the DataFrame to confirm the changes in the two steps above

df_age_aggregated.head()

#### Write the cleaned dataframe to a new CSV file and save the output in an S3 bucket

# Instantiate boto3 by providing access and secret keys
client = boto3.client('s3', aws_access_key_id='AKIATNJHRXAPUA4DIFER', aws_secret_access_key="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")

# Provide the name of the s3 bucket
bucket = "2207-17-fibre-competitive-intensity-model-b"

# Convert df to CSV
df_age_aggregated.to_csv('sa_2011_age_aggregated_decimal_final.csv', index = False)

# Save CSV file to S3 bucket
client.upload_file("sa_2011_age_aggregated_decimal_final.csv", Bucket=bucket, Key="sa_2011_age_aggregated_decimal_final.csv")
