==> notebooks/2.8-prepare-ward-level-data-for-modelling.ipynb <==
# Prepare Ward-Level Data for Modelling

# Import the necessary libraries
import numpy as np
import pandas as pd
import boto3

# Display all columns in the dataframes
pd.set_option("display.max_columns", None)

# Instantiate boto3 by providing access and secrete keys
client = boto3.client('s3', aws_access_key_id='AKIATNJHRXAPUA4DIFER', aws_secret_access_key="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")

# S3 bucket name
bucket = "2207-17-fibre-competitive-intensity-model-b"

# Generate a file path to S3 bucket
file_path_demo = 'https://2207-17-fibre-competitive-intensity-model-b.s3.eu-west-1.amazonaws.com/Preprocessed+Data/ward-level-demography.csv'
file_path_uptake_rate = 'https://2207-17-fibre-competitive-intensity-model-b.s3.eu-west-1.amazonaws.com/Preprocessed+Data/SA-uptake-rate-ward-level.csv'

# Load the datasets
df_demography = pd.read_csv(file_path_demo)
df_uptake_rate = pd.read_csv(file_path_uptake_rate)

# Preview the demography data
df_demography.head()

# Preview the uptake rate data
df_uptake_rate.head()

# Drop cloumns that are not required
df_demography = df_demography.drop('Unnamed: 0', axis=1)
df_uptake_rate = df_uptake_rate.drop('Unnamed: 0', axis=1)

The datasets are clean and requires no further preprocessing. Next is to merge the datasets

# Merge the datasets
df_ward_data = pd.merge(df_demography, df_uptake_rate, on=['ward_code','ward'])

# Preview the merged data
df_ward_data

# "total_pop" is thesame with population, so one of them wiil be dropped
df_ward_data = df_ward_data.drop('total_pop', axis=1)

# Export the dataset for modelling
df_ward_data.to_csv('ward-data-cleaned.csv', index=False)

# Upload the file to the S3 bucket
client.upload_file("ward-data-cleaned.csv", Bucket=bucket, Key="Data for Modeling/ward-data-cleaned.csv")
