==> notebooks/2.9-prepare-and-link-admin-boundaries-for-streamlit-and-model-predictions.ipynb <==
# Import the necessary libraries
import pandas as pd
import geopandas as gpd
import boto3

# Instantiate boto3 by providing access and secrete keys
client = boto3.client('s3', aws_access_key_id='AKIATNJHRXAPUA4DIFER', aws_secret_access_key="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")

# S3 bucket name
bucket = "2207-17-fibre-competitive-intensity-model-b"

# Create a file path to all shapefiles of interst
ward_path = "https://2207-17-fibre-competitive-intensity-model-b.s3.eu-west-1.amazonaws.com/Raw+Data/2011+Wards+Administrative+boundaries/Wards2011.shp"

# Load the datasets
# This line of code may not work as geopandas requires all shapefiles to be in thesame folder so as to interract with other files
# needed. But the file path from S3 bucket may most likely not give that access to geopandas. The solution may be to run the
# shapefiles locally or to run their geojson format from S3 bucket
gdf_ward = gpd.read_file('https://2207-17-fibre-competitive-intensity-model-b.s3.eu-west-1.amazonaws.com/Streamlit+Data/ward.geojson')

# Preview the datasets
gdf_ward.head()

# Extract geometry for all levels of admin boundary
gdf_ward_new = gdf_ward[['PROVINCE','MUNICNAME','WARD_ID','geometry']]

# Dissolve to get the polygons for municipality and province level
munic_geo = gdf_ward_new.dissolve(by='MUNICNAME').reset_index()
province_geo = gdf_ward_new.dissolve(by='PROVINCE').reset_index()

# Get the admin boundaries info on ward level
admin_bound = gdf_ward[['PROVINCE','MUNICNAME','WARD_ID','WARD_POP','Area']]

# Reproject the province dataframe to a geographical CRS
province_crs = province_geo.to_crs(4326)

# Get the latitudes and longitudes of each province
province_crs['lat'] = province_crs.centroid.y
province_crs['long'] = province_crs.centroid.x

# Get the province and their respective latitudes and longitudes 
province_crs = province_crs[['PROVINCE','lat','long']]

# Reproject the dissolved municipality dataframe to a geographical CRS
munic_crs = munic_geo.to_crs(4326)

# Get the latitudes and longitudes of each municipality
munic_crs['lat'] = munic_crs.centroid.y
munic_crs['long'] = munic_crs.centroid.x

# Get the province and their respective latitudes and longitudes 
munic_crs = munic_crs[['MUNICNAME','lat','long']]

# Export province, municipality and ward shapefiles in a geojson format
province_geo.to_file('prov.geojson', driver='GeoJSON')
munic_geo.to_file('muni.geojson', driver='GeoJSON')
gdf_ward.to_file('ward.geojson', driver='GeoJSON')
province_crs.to_csv('province-cord.csv')
munic_crs.to_csv('munic-cord.csv')
admin_bound.to_csv('admin_boundaries_2011.csv')

# Upload the files to the S3 bucket
client.upload_file("prov.geojson", Bucket=bucket, Key="Streamlit Data/prov.geojson")
client.upload_file("muni.geojson", Bucket=bucket, Key="Streamlit Data/muni.geojson")
client.upload_file("ward.geojson", Bucket=bucket, Key="Streamlit Data/ward.geojson")
client.upload_file("province-cord.csv", Bucket=bucket, Key="Streamlit Data/province-cord.csv")
client.upload_file("munic-cord.csv", Bucket=bucket, Key="Streamlit Data/munic-cord.csv")
client.upload_file("admin_boundaries_2011.csv", Bucket=bucket, Key="Preprocessed Data/admin_boundaries_2011.csv")
